# Add test binary (speed up the test)