--- a/testsuite
+++ b/testsuite
@@ -36,7 +36,7 @@ _run() {
     shift 2
     port=$1
     shift 3
-    "$WD/coc" -t stderr $1 $2 $3 $4 $5 $6 $7 $8 $9 -- "$WD/tcpcontest" "$host" "$port"
+    "$WD/coc" --test-mode -t stderr $1 $2 $3 $4 $5 $6 $7 $8 $9 -- "$WD/tcpcontest" "$host" "$port"
 }
 
 _header() {