This is another version of #2.

I added test binary to coc and speed up the test.
Test binary actually blocks connections even if the result is ALLOW.
This is an option for coc development. (for testsuite)

# Motivation

- CIDR notation support (#1) increased the number of tests, so I wanted to speed up the test.
- I wanted to write a test without worrying about connecting to real IP.

# Difference from #2

I prepared a test binary and used it with coc's command line option(--test-mode).

- Pros
  - It is not affected by environment variables.
- Cons
  - Increase Makefile rule. (for test binary)

# Benchmark

On my computer.  With master HEAD(21b859ac70).
132 sec -> 3 sec.

## Without --test-mode option

```
$ time make test > /dev/null

real    2m12.778s
user    0m2.743s
sys     0m0.674s
```

## With --test-mode option

```
$ time make test > /dev/null

real    0m2.808s
user    0m2.738s
sys     0m0.650s
```


# Test results

## Without --test-mode option

The behavior of coc has not changed.

```
% ./coc -l allow -- ./tcpcontest 8.8.8.8 53
2019-01-28T11:37:45 ALLOW connection to 8.8.8.8:53
connect to 8.8.8.8 is OK

% ./coc -l allow -a "*" -- ./tcpcontest 8.8.8.8 53
2019-01-28T11:38:04 ALLOW connection to 8.8.8.8:53
connect to 8.8.8.8 is OK
```

see also
https://travis-ci.org/tgg/connect-or-cut/builds/436477149
> Total time 1 hr 59 sec

## With --test-mode option

The message is ALLOW, but it does not actually connect.

```
% ./coc --test-mode -l allow -- ./tcpcontest 8.8.8.8 53
2019-01-28T11:38:30 ALLOW connection to 8.8.8.8:53
connect to 8.8.8.8 is KO: errno is 13 (Permission denied)

% ./coc --test-mode -l allow -a "*" -- ./tcpcontest 8.8.8.8 53
2019-01-28T11:38:57 ALLOW connection to 8.8.8.8:53
connect to 8.8.8.8 is KO: errno is 13 (Permission denied)
```

see also
https://travis-ci.org/tgg/connect-or-cut/builds/485226270
> Total time 1 min 42 sec