@saito-tom 2022-01-31
I'll close the pull request since it doesn't seem to be actively being developed these days.
If you need this, please incorporate the changes.